// galaxyui/src/app/resources/respositories/repository.ts
export class Repository {
    id: number;
    name: string;
    original_name: string;
    description: string;
    is_enabled: boolean;
    provider_namespace: any;
    url: string;
    github_user: string;
    github_repo: string;
    import_branch: string;
    commit: string;
    commit_message: string;
    commit_url: string;
    commit_created: string;
    stargazers_count: number;
    watchers_count: number;
    forks_count: number;
    open_issues_count: number;
    summary_fields: any;
    related: any;
}